Adjustment to module imports due to renaming of the directory for metadata field types
