+++ client/src/app/metadata-schema-form/metadata-schema-form.module.ts
@@ -5,23 +5,23 @@ import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {MatTabsModule} from '@angular/material/tabs';
 import {RouterModule} from '@angular/router';
 import {MetadataFormComponent} from './metadata-form/metadata-form.component';
-import {InputComponent} from './input-types/input/input.component';
+import {InputComponent} from './metadata-field-types/input/input.component';
 import {VfInputComponent} from './custom/vf-input/vf-input.component';
 import {MetadataFormItemDirective} from './metadata-form-item.directive';
-import {TextListInputComponent} from './input-types/text-list-input/text-list-input.component';
-import {TextAreaComponent} from './input-types/text-area/text-area.component';
+import {TextListInputComponent} from './metadata-field-types/text-list-input/text-list-input.component';
+import {TextAreaComponent} from './metadata-field-types/text-area/text-area.component';
 import {VfAsteriskComponent} from './custom/vf-asterisk/vf-asterisk.component';
-import {DateInputComponent} from './input-types/date-input/date-input.component';
-import {OntologyInputComponent} from './input-types/ontology-input/ontology-input.component';
+import {DateInputComponent} from './metadata-field-types/date-input/date-input.component';
+import {OntologyInputComponent} from './metadata-field-types/ontology-input/ontology-input.component';
 import {MultipleSelectComponent} from './custom/multiple-select/multiple-select.component';
 import {MetadataFieldComponent} from './metadata-field/metadata-field.component';
 import {SelectComponent} from './custom/select/select.component';
-import {OntologyListInputComponent} from './input-types/ontology-list-input/ontology-list-input.component';
-import {EnumListInputComponent} from './input-types/enum-list-input/enum-list-input.component';
+import {OntologyListInputComponent} from './metadata-field-types/ontology-list-input/ontology-list-input.component';
+import {EnumListInputComponent} from './metadata-field-types/enum-list-input/enum-list-input.component';
 import {MaterialModule} from '../material.module';
-import {OntologyBaseComponent} from './input-types/ontology-base/ontology-base.component';
+import {OntologyBaseComponent} from './metadata-field-types/ontology-base/ontology-base.component';
 import { MetadataFormItemComponent } from './metadata-form-item/metadata-form-item.component';
-import { BaseInputComponent } from './input-types/base-input/base-input.component';
+import { BaseInputComponent } from './metadata-field-types/base-input/base-input.component';
 
 
 @NgModule({
